Moved upload to component directory
