+++ logicle/app/chat/components/ChatInput.tsx
@@ -6,7 +6,7 @@ import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import 'react-circular-progressbar/dist/styles.css'
 import { Attachment } from '@/types/chat'
-import { Upload, UploadList } from './Upload'
+import { Upload, UploadList } from '../../../components/app/upload'
 import { post } from '@/lib/fetch'
 import { File, InsertableFile } from '@/types/db'
 import toast from 'react-hot-toast'
